@SuperSandro2000 2023-01-30
> webservers usually only enable them if they are advertised through the Accept-Encoding header.

They must only send what is written in Accept-Encoding. Thats why in addition to zstd supporting at least brotli and usually gzip, too is necessary for a normal webserver. Most browsers don't have adopted zstd yet or only behind a flag. Not that relevant for us here since nix support brotli for a while.