@NickCao 2023-01-30
My earlier attempt at this: https://github.com/NixOS/nix/pull/6694