@Mic92 2023-01-31
Great @NickCao has implemented the other approach in https://github.com/NixOS/nix/pull/6694, where nix offloads decompressing to libcurl. This means both variants are implemented and you just need to merge the one that you prefer.